pyproject.toml
[project]
name = "pydantic-data-quality"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.3",
    "pandera>=0.26.1",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]

[tool.taskipy.tasks]
format = """
isort .
black .
"""

kill = "kill -9 $(lsof -t -i :8000)"
test = "pytest -v"
run = """
python3 app/main.py
"""

doc = "mkdocs serve"
